apps/job/serializers.py
from rest_framework import serializers
from django.utils.crypto import get_random_string
from django.db import transaction, IntegrityError
from rest_framework.authtoken.serializers import AuthTokenSerializer
from rest_framework import exceptions
from rest_framework import status

