requirements/base.txt
Django==3.0.8
django-allauth==0.42.0
django-appconf==1.0.4
django-celery-beat==2.0.0
django-compressor==2.4
django-crispy-forms==1.9.2
django-environ==0.4.5
django-model-utils==4.0.0
django-redis==4.12.1
django-timezone-field==4.0

# Django REST Framework
djangorestframework==3.11.1

# We have debug-toolbar also in production in some scenarios
django-debug-toolbar==2.2

celery==4.4.7
flower==0.9.5

whitenoise==5.2.0

requests==2.24.0
requests-oauthlib==1.3.0

amqp==2.6.1
argon2-cffi==20.1.0
asgiref==3.2.10
billiard==3.6.3.0
certifi==2020.6.20
cffi==1.14.2
chardet==3.0.4
cryptography==3.0
defusedxml==0.6.0
ecdsa==0.14.1
humanize==2.6.0
idna==2.10
importlib-metadata==1.7.0
kombu==4.6.11
oauthlib==3.1.0
Pillow==7.2.0
prometheus-client==0.8.0
pyasn1==0.4.8
pycparser==2.20
PyJWT==1.7.1
python-crontab==2.5.1
python-dateutil==2.8.1
python-slugify==4.0.1
python3-openid==3.2.0
pytz==2020.1
rcssmin==1.0.6
redis==3.5.3
rjsmin==1.1.0
rsa==4.6
six==1.15.0
sqlparse==0.3.1
text-unidecode==1.3
tornado==6.0.4
urllib3==1.25.10
vine==1.3.0
zipp==3.1.0

# auth0 stuff
social-auth-app-django==3.4.0
social-auth-core==3.3.3
python-jose==3.2.0

# chargebee
chargebee==2.7.4
